Autorização concluido

+++ app/Http/Controllers/ProjectController.php
@@ -31,7 +31,7 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        return $this->repository->all();
+        return $this->repository->findWhere(['owner_id' => \Authorizer::getResourceOwnerId()]);
     }
 
     /**
@@ -53,6 +53,10 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
+        if ($this->checkProjectPermissions($id) == false) {
+            return ['error' => 'Acesso Negado'];
+        }
+
         return $this->repository->find($id);
     }
 
@@ -65,6 +69,9 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if ($this->checkProjectPermissions($id) == false) {
+            return ['error' => 'Acesso Negado'];
+        }
         return $this->service->update($request->all(), $id);
     }
 
@@ -76,6 +83,32 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
+        if ($this->checkProjectPermissions($id) == false) {
+            return ['error' => 'Acesso Negado'];
+        }
         $this->repository->delete($id);
+    }
+
+    private function checkProjectOwner($projectId)
+    {
+        $userId = \Authorizer::getResourceOwnerId();
+
+        return $this->repository->isOwner($projectId, $userId);
+    }
+
+    private function checkProjectMember($projectId)
+    {
+        $userId = \Authorizer::getResourceOwnerId();
+
+        return $this->repository->hasMember($projectId, $userId);
+    }
+
+    private function checkProjectPermissions($projectId)
+    {
+        if ($this->checkProjectOwner($projectId) or $this->checkProjectMember($projectId)) {
+            return true;
+        }
+
+        return false;
     }
 }
